LayoutTests/fast/sandbox/ios/sandbox-mach-lookup-attachment-element.html

<!DOCTYPE html><!-- webkit-test-runner [ enableAttachmentElement=true ] -->
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description('Test that the WebContent process has mach lookup access to frontboard and icon services when the attachment element is enabled');

if (window.internals) {
    shouldBeTrue("internals.hasSandboxMachLookupAccessToGlobalName('com.apple.WebKit.WebContent', 'com.apple.frontboard.systemappservices')");
    shouldBeTrue("internals.hasSandboxMachLookupAccessToGlobalName('com.apple.WebKit.WebContent', 'com.apple.iconservices')");
}
</script>
</head>
<body>
</body>
